Weston Performance Fix
This fixes the issue where Weston wasn't presenting at the rate
it was expected to. In particular, the RDP backend would always
take 16ms to complete a frame after being sent to the host. Upon
frame completion, Weston would calculat the next frame to present
based on the monitor refresh and target 9ms ahead of that. At 64hz
this ended up having Weston compositor wait for 9ms for the next
frame and that 9ms would add up to the RDP monitor 16ms simulated
VSYNC, resulting in an effective presentation rate of 1/25ms = 40fps.

The VSYNC emulation is now improved and the RDP backend keeps the
frame completion within the total expected refresh, so at 64hz
the frame are kept a total of 16ms apart.

Also updated the VSYNC emulation code to adapt the RDP monitor
refresh to what is specified through the environment variable
WESTON_RDP_MONITOR_REFRESH_RATE.