rdp: Remove westonRect from backend structures
Change this to a local variable now that the two functions that should be
using it are combined.

This commit intentionally introduces a bug that will be fixed in a follow
up commit. Outputs may not be properly moved into place immediately
after creation.

In the future output creation and positioning will happen in the front end,
and this will no longer be a problem.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>