Fall back to weston.ini keymap if no mapping is found (#144)
* rdp backend: do not force us layout

Do not force "us" layout when no xkb mapping is found. Weston will this
way fallback to the keymap defined in compositor xkb_names which falls
back to "us" if no layout is defined in weston.ini.

* rdp backend: reset default keymap if mapping fails

Reload default global compositor keymap when no xkb mapping is found for
current keyboard layout. This will default to the keymap in weston.ini
or to "us" if no keymap is defined there. Should allow to use custom KLC
layouts for keymaps that exist in Linux but have no correspondence in
Windows (e.g. altgr-intl). This should also partially fix #173. Still
need a way to export weston.ini to the user.